For many critical applications where URBI has to handle "lists" of objects created at runtime (like bots in a video game), it is necessary to be able to have objects refered to by names containing variable parts. This feature is added via the hash map mecanism already existing:
class x;
obj[1] = new x;
obj["myname"] = new x;
obj[45]["hello"] = new x;

The way it is done is again via a hack in the nameresolution function and the
parser. In k1.5, this will be fixed with ageneric nameresolution algorithm.

NB: obj[1] is not refering to an underlying ulist of objects, and evaluating 
obj alone without the array index will fail.

	add support for objects hash table 
	* src/parser/bison/utoken.l: add POINT token ('.').
	* src/parser/bison/ugrammar.y: add support for prefix[...].suffix[...]
        constructs in grammar (should be generalized in 1.5).
	* src/uvariablename.cc: extend buildFullname function and fix bugs in
        name resolution algorithm.
	* src/uvariablename.h: add index_obj array.

git-svn-id: file:///srv/svn/kernel1/trunk@284 1e37f2df-8004-0410-ba96-adcf012cb36e