Add a yield when evaluating a argumentless Urbi function
Arguments evaluation cause a yield to happen. However, function which
do not have arguments can go into an infinite loop without ever yielding.

Only code written in Urbi yields, as it is expected for a primitive written
in C++ and not calling back Urbi code not to enter an infinite (or even
long timed) loop.

	* src/runner/runner.cc: Yield before starting to execute Urbi
	code.