Look for the function to be called before evaluating the arguments
This has no influence on the test suite as no test relied on such an
unspecified behaviour. However, if we ever plan to implement call messages,
this order of evaluation is sensible. One could also agree that it makes
no sense to evaluate arguments that will never be used in a call as an
error will be raised just thereafter.

The only time where this could have any importance is if an argument
evaluation happened to set/update/remove the slot corresponding to the
method called. And we do not want to condone this behaviour anyway.

	* src/runner/runner.cc (operator() (ast::Call)): Look for the
	executable slot before evaluating the arguments.