Use subdir-objects.
In order to improve make parallelism, we don't use on Makefile.am per
directory, but rather several *.mk file per directory included from the
main Makefile.am.  Unfortunately the consequence is that if there are
several directories with the same file name, we have conflicts on the object
files, unless we use subdir-objects (which means to create sub/foo.o instead
of foo.o from sub/foo.c).

But then we have a problem: we must not use Make variables in source
file names, because it breaks Automake's implementation of "include" and
the creation of .deps files.  So, this is ugly, but we have to inline
in the main Makefile the path to the subdir sources instead of relying
on Make variables.

This is really ugly, and maybe we'll find something nicer, time permiting.

	* configure.ac, Makefile.am: Now use lib/libport as a real subdir.
	* src/Makefile.am, src/kernel.mk: Adjust.
	Hard code the path to the urbi-sdk-remote files.
	No longer compile libport.la here.