Fix weird behavior when using network
When receiving several lines in one time from the network, the kernel sometimes
died. For each line, it put the corresponding ast into the main Nary, and run a
fresh runner on it, even if the execution of the previous line was not
finished, and the Nary cleaned. Therefore some command was runned several
times, and the Nary was cleaned while some runners was on it.

If there is already a runner on the main Nary, new commands are now just pushed
at the end, the current runner will handle them.

	* src/uconnection.cc: In received_(), if the Nary active_command_ is not
	  empty, there is already a Runner on it, so do not launch an other one.
	* include/kernel/uconnection.hh: obstructed becomes a local variable of
	  received_().