Protect all the changes of valp.
My current approach was too naive: we need to be able to
change the start conditions to avoid catching ( and so
forth in strings or comments.  My previous patch, which
does not run the actions at all, does not change start
conditions, so it fails to protect strings etc.

Now that I plan to run the actions for real, we do not
want to grow valp when its value is not used.  So use
a macro for all the assignments to valp.

	* src/parser/utoken.l (SET_VAL, GROW_VAL, APPEND_VAL, STEP, ERROR):
	New.
	Use them.
	(LINES): Take the number of lines as arg.
	Also, report the guilty escapes.
	* tests-local.mk (TFAIL_TESTS): Add symbols.chk.