Let the scanner do the whole prescanning.
Before, the scanner was used to return the interesting
characters (braces and terminators), and uqueue was doing
the bookkeeping.  Now, the scan has two modes: the regular
mode, and the prescanning mode.  In the latter mode, it
scan up to the next terminator (or maybe brace-balance
error).  UQueue simply consults it to the get the length
of the next buffer chunk.

	* src/parser/utoken.hh: New.
	* src/parser/parser.mk: Adjust.
	* src/parser/utoken.l (YY_USER_ACTION): No longer match
	prescanned characters, let the actions do that.
	(PRESKIP): Rename as...
	(PRE_SKIP): these.
	(PRE_PUSH, PRE_POP, PRE_RETURN, pre_closers): New.
	Use them.
	(prescan): New.
	* src/uqueue.cc (pop_command): Use it.