Even when retrieved from C++ code, self may need to be executed
This patch and the previous ones fix an horrible (but funny after analysis)
bug, due to the fact that any member in Urbi may be a value or a parameterless
function returning a value.

The bug revealed itself in a lobby as:

Object.a = 1; a = 2; a;
==> 1

Where did the 2 go? It went on lobby.self which was not an object
but a primitive.

lobby.getSlot("self");
==> <primitive>
lobby.getSlot("self").a;
==> 2

	* src/object/object.hh (slot_update): Take a runner as parameter
	to evaluate "self".
	* src/object/object.cc: Ditto.
	* src/object/object-class.cc, src/uobject.cc: Use the new interface.
	* tests-local.mk: 2.x/basic.chk is now a PASS.