Simplify the definition of primitive_class.
Its previous definition was complex because we meant to define and
use primitive_type, which must be done after rObject was defined (since
it is used in primitive_type) and before primitive_class is defined
(since it uses primitive_type).

We no longer define primitive_type, which allows to vastly simplify.

	* src/object/fwd.hh (APPLY_ON_ALL_PRIMITIVES_BUT_OBJECT_AND_PRIMITIVE):
	Remove, fused into...
	(APPLY_ON_ALL_PRIMITIVES_BUT_OBJECT): this.
	(primitive_type): Remove.
	* src/object/atom.hh (primitive_traits::type): Adjust.