Add clone/set/+=/empty to String.
We have a problem here: these function should be available for all the
Atoms, or for none...  But if we keep it, it should be factored.  Either
a dirty way (macros), or a nice way (a common parent to all atoms?
It might not be possible because of C++ typing issues).

	* src/object/string-class.cc (String.set, String.clone): New.
	* share/urbi/string.u (String.+=, String.empty): New.