Use ast::String.
This is basically a revert of f976c1e5b31de7fd8c92d5ea1461c79dcb2c19a9.
It is painful to use dynamic structures (object::Object) within the AST.
It does factor a lot of code, but (i) it becomes painful to use now
that printing an object::Object requires a Runner (therefore YYDEBUG is
broken, and has been broken for a long time), (ii) it augments the
dependencies between different modules (ast needs object), and anyway
(iii) it goes in the wrong future direction where we want a faithful
AST desugared in later steps.

Therefore all of ast::Object is scheduled for removal.

	* src/ast/ast.yml (String): New, again.
	* src/parser/ugrammar.y: Use it.
	* src/runner/runner.cc, src/runner/runner.hh: Adjust.