Do not enqueue calls in case of a frozen job
If a job is frozen, it will enqueue a potentially long list of calls to

  Job::yield()
  Scheduler::resume_scheduler() <-------+
  Scheduler::switch_back()              |
  Job::yield_until_things_changed()  ---+

This patch reorganizes the resume_scheduler() by integrating the
switch_back() code, without yield_until_things_changed(), into the
method, and prevents the loop.

Also, it is slightly more efficient as it won't check for possible
stack overflow if we are either not restarting the job or climbing
up the call stack because of an exception (meaning that the stack
will get more free space).

	* src/scheduler/scheduler.hh (switch_back): Remove.
	* src/scheduler/scheduler.cc: Ditto.
	(resume_scheduler): Reorganize and loop until job needs
	to be restarted.