Use an ast::Scope as function body
A function body was previously made of a Nary. It means that the
non-interruptible status had to be saved around a Nary execution, which
is already quite a mess because of the multiple exit paths and ways of
handling results and errors (depending on whether we are at top-level or
not).

This patch makes the function body an ast::Scope, and removes the inner
scope creation when running a function, as one will be created
automatically.

As a consequence, locals.slotNames within a function will not show the
arguments, but nothing relies on it and this is cleaner anyway.

	* src/ast/ast.yml: Function body is now a ast::Scope.
	* src/parser/ugrammar.y: Ditto.
	* src/runner/runner.cc: Do not create an inner scope as it will
	be created as part of going over the function body.