Desugar "i++" as "(i = i + 1) - 1".
Previously, ++, += and the others were simple messages sent to their
argument.  As a result, "a++" was actually having a side effect on the
value of "a" (i.e., what was really run is "1++") and any other object
which had a copy of "a" also had it value changed.  In Urbi instead of
word:

	var a = 1;
	var b = a;
	a++;
	assert(b == 2);

To change this, we now desugar "a++" as "(a += 1) - 1", and we also desugar
"a += 1" into "a = a + 1".  As a side-effect, now these constructs
only apply on lvalues, not arbitrary exps.

	* src/parser/tweast.hh, src/parser/tweast.hxx, src/parser/tweast.cc:
	Now also support ast::Call metavariables.
	It seems redundant since we already have ast::Exp from which ast::Call
	derives, but (i) we really want to get an ast::Call when using it as
	an lvalue, and (ii) anyway it helps getting rid of C++ problems
	when we try to store an ast::Call inside a ast::Exp: the match is
	not perfect and therefore we have (according to the C++ std) an
	ambiguity with using template<T> append_(const T&).
	(move_metavars_): Remove dead declaration.
	* src/parser/utoken.l (_call): Recognize it.
	(++, --, +=, -=, *=, /=): Now simple tokens without semantic value.
	* src/parser/ugrammar.y: Adjust the token definitions.
	(DESUGAR): Rename as...
	(DESUGAR_): this.
	(DESUGAR): New, works on $$ directly.
	Use it.
	(++, --, +=, -=, *=, /=): Desugar into assignments.
	* share/urbi/urbi.u (Scope.atExit): Don't use += on an expr, we
	want an lvalue.
	* tests-local.mk: We pass 2.x/intops.chk.