Handle chained messages on groups.
This code:

var g = Group.new << ...; // Create a new group with members
g.foo.bar;

must apply foo on every member of the group, and then bar on every result of
foo. This is needed to have 'group.value = 42' behave as expected, for example.
This beahvior is obtained really simply by making messages applied on group
return a group containing the result.

    * share/urbi/group.u
    (add): Return self, to enable chaining insertions.
    (asString): Print groups relevantly.
    (fallback): Return a group to distribute chained messages as expected