Temporary TrajectoryGenerator 'fix'.
Since writting a value to a variable can create a new trajectory generator
if the speed property is set, TrajectoryGenerator::run calls changeSlot,
which bypasses the updateHook. As a consequence, it can't be used on
UObject variables. This patch disables the bypass. It works as long as no
property is used.
We need a stackable hooking mechanism.

	* share/urbi/trajectory-generator.u(run): Call updateSlot instead
	  of changeSlot.