Remove the dependency from the scheduler to the userver.
We already have a dependency from the userver to the scheduler, as the
former owns an instance of the latter.

In order to avoid dependency loops, the server gets a functional object
to call to get the time in its constructor. This way, it could be used
independently from an Urbi server.

	* src/scheduler/scheduler.hh, src/scheduler/scheduler.hxx: Take
	and store a funcitonal object in the constructor.
	* src/scheduler/scheduler.cc (work, execute_round): Use it.
	* userver.cc (UServer): When creating a scheduler, give it a
	reference to our getTime member function.