Use n-ary add/remove in groups instead of two sets of functions.
This does not work too well because in order to remove members from
an heterogeneous list, we need to compare them with the guy to be removed.
I use "==" to this end, and '1 == "2"' fails (an error) because they don't
have the same type.

I think this is a misfeature, and rather *when* we want type-checking
==, which should say so first.  But one could argue the converse: issue
special type-checking code first if you want to accept heterogeneous
comparisons.

WDYT?

	* share/urbi/group.u (Group.addMembers, Group.removeMembers): Remove.
	(Group.add, Group.remove): n-ary functions.
	* tests-local.mk: We don't pass groups.chk.