Parametric ASTs.
Implement a new scheme for concrete syntax support.  Instead of
parsing the TWEAST at each use, parse it with meta-variables in
it, and in a second step, fill in the blanks.  This allows to keep
a copy of the parametric AST, and therefore to parse only once.

More documentation will come later.

	* src/ast/ast.yml (MetaExp): New.
	* src/ast/parametric-ast.hh, src/ast/parametric-ast.hxx,
	* src/ast/parametric-ast.cc: New.
	* src/ast/local.mk: Adjust.
	* src/parser/utoken.l (%exp:): New token.
	* src/parser/ugrammar.y: Use it to desugar constructs that use
	only meta-variables of type Exp*.