Make methodToFunction more powerfull.
Before this patch, one had to know the method to make it a function. It was
possible to make a function of Object.asString for instance, but applying it
on an object would yield the Object.asString behavior, bypassing the potential
object's specific asString.

Only the method name is now needed, and the right method is called on any
object.

    * share/urbi/urbi.u (methodToFunction): Perform lookup on the target to find
    the right method.