Stop storing null pointers in exps_type.
I am trying to use ptr_containers for it, and it is really
much more complex to allow empty pointers in theses containers.
I feel it is safer to still use ptr_containers, and get rid
of null pointers.  The problem here is that we are wasting
space with these many identical objects: a singleton would
be much more useful, but the tool set is not ready yet.
Besides, it might be better in the end to use another approach
with a local hierarchy for Call: Calls with a target, and
Calls without.

	* src/ast/ast.yml (Implicit): New.
	(Call): Don't print them.
	* src/parser/ugrammar.y, src/runner/runner.cc: Adjust.