Make target invisible from the urbi side.
Previously, target was a Scope method, and was invoked to determine the actual
target of unqualified calls. Every call thus entailed a lookup for target. It
was implemented this way to enable user - and us developers - to easily
override it, thus changing the target of unqualified calls in some scope. Yet,
overriding it would be quite a hack, and this kind of features will disappear
anyway when we'll switch to static binding. Target uses are now hardcoded in
C++, thus saving a lot of time.

This patch yields a 25% speedup.

    * src/object/scope-class.cc
    (scope_class_tryTarget): Remove.
    (scope_class_locateSlot): Call C++ target directly.
    * src/runner/runner.cc (Runner::target): Call C++ target directly.
    * src/parser/ugrammar.y: Use Scope.target instead of target alone to
    speedup lookups.