No longer support operator as regular identifiers.
They need to be quoted with ''.  This is because it does not work
for all the operators (unary ones cannot be supported) and adds too
much constraints to the parser.

I am ambivalent wrt this patch, because it means that we can no longer
write this like:

	do 1 { * 2 }

etc.  But was it really so cool?

Also, should we support "operator*" to stand for '*'.  My guts say
we should, for obvious reasons.

	* src/parser/ugrammar.y (id): Remove all the operators from it.
	"new" is the only magic supported keyword/identifier.
	* tests: Update.