Use Sequences instead of Tuples with Boost.Preprocessor.
Tuples are limited to 25 elements, Sequences go up to 256, and are
faster and easier to use.

I have also disable the MSVC 2005 SP1 tricks.  It should be easy
to import them again, but let's do that afterwards.  Reading this
patch, which remove the support, should make it easy to reenable it.
In particular by putting the typedef thingies in the deeper macros
(VISITOR_VISIT_NODE_ and the like).

	* dev/ast.py (visit): Prepare a sequence instead of a tuple.
	If 'const' is not a Boolean, then let it be a macro name.
	* dev/ast-visitor-gen: Define many macro using Boost.Preprocessor
	to factor the declaration of visit methods.
	Use Sequences instead of Tuples.
	* dev/ast-default-visitor-gen: Adjust to using them.
	* src/ast/visitor.hxx, src/runner/runner.hh: Adjust.