Switch to static, lexical local variables binding.
This patch removes all the 'target' stuff that used to determine at runtime who
was the target of the message. This is now determined by static, lexical
binding. This task is completely modular and stand alone: it works by rewriting
the parsed ast before evaluating it, to retarget messages. After this pass,
all messages are to be applied on locals, which greatly simplifies the runner
jobs: no need anymore to explore hierarchies to determine which object is the
target.

The static binding also handles full implicit read/write lexical closure: no
need for capture, makeClosure, ... anymore.

The rewriting looks like this:

{
  var closure;
  function (arg)
  {
    var local;
    arg;
    local;
    closure;
    member;
  }
};
[00041604] function (arg) {
  setSlot("local", getSlot("self") . nil);
  arg;
  local;
  getSlot("code") . context . closure;
  getSlot("self") . member;
}

This is also the first half of the work to get rid of the scope-are-objects
thing.

On the other hand, the way to find global symbols is now through self again,
which need to be discussed.

    * src/binder/binder.cc,
    * src/binder/binder.hh,
    * src/binder/binder.hxx,
    * src/binder/fwd.hh,
    * src/binder/local.mk: New. Local variables binder visitor.
    * src/Makefile.am: Build src/binder/*.
    * share/urbi/dictionary.u,
    * share/urbi/group.u,
    * share/urbi/k1.u,
    * share/urbi/lazy.u,
    * share/urbi/list.u,
    * share/urbi/tuple.u,
    * share/urbi/urbi.u: Adapt. Essentially simplify code to use implicit
    closures.
    * share/urbi/urbi.u: Remove all closure and scoping code.
    * src/object/lazy.cc: Bind generated function so as they look for variables
    in the context.
    * src/object/object.cc,
    * src/object/object.hh: Remove all target related code.
    * src/object/root-classes.cc: Make Object inherit from Global.
    * src/object/scope-class.cc: Remove all now useless *Slot overloads.
    * src/uconnection.cc: Bind received code.
    * src/object/system-class.cc: Bind loaded files.
    * src/runner/runner.cc,
    * src/runner/runner.hh,
    * src/runner/runner.hxx: Remove target lookups. Simply send all implicit
    messages to the locals.
    * tests-local.mk: Reorganize. Mark some tests as TFAIL. The binder really
    breaks only one test (2.x/timeshift2.chk). The other TFAIL are location or
    other trivial issues that I need to examine.
    * tests: Update test dependency.