Implement synchronous tag stopping.
The logic of tag stopping and blocking has been totally revamped.
Now, instead of executing a new round in the scheduler when one or more
tags have been stopped, each job is given the opportunity to compute how
far it has to rewind its execution stack to cancel all the stop or
blocked tags.

Also, trying to execute a statement tagged with a blocked tag will skip
this statement, instead of interrupting it right at the beginning by
throwing an exception.

It allowed some simplification in the scheduler workflow, and also
alleviated the need for the blocked() method in jobs.

The changes would be too long to describe file by file, the above
description seems much more useful to me.