First stab at reactivating modifier support.
More will be needed, in particular because of the ongoing changes
on the status of the variable (field vs. local).

The former implementation used to rely on plain objects (| ... |)
whose support was discontinued.  This implementation relies on the
Dictionaries instead.  But to avoid introducing syntactic support
for Dictionaries, we now may grow tweast during the parsing.

	* share/urbi/trajectory-generator.u: Convert to using "args" as
	a dictionary, not a plain object.
	* src/parser/ast-factory.hh, src/parser/ast-factory.cc
	(ast_slot_change): Take a modifier, and use it.
	* src/parser/ugrammar.y (modifier, modifiers, modifiers.1): new.
	Rely on tweast for their support.
	Use their values.