Create classes before initializing them.
Some classes need other primitive classes to initialize themselves. Thus, first
create all the classes, then initialize them all.

Also add a systematic copy constructor to primitive class, and tweak the clone
method to call it when, on the Urbi side, one clones an instance of a primitive,
not the primitive class itself. That is:

Float.clone;
[12345678] 0
42.clone;
[12345678] 42

    * src/object/cxx-object.cc,
    * src/object/cxx-object.hh,
    * src/object/cxx-object.hxx
    (Initializer::create): New. Create all the classes.
    (Initializer::Initializer): Now take the rObject to fill with the created
    class.
    * src/object/code-class.cc,
    * src/object/code-class.hh,
    * src/object/dictionary-class.cc,
    * src/object/dictionary-class.hh: Add "copy" constructor. Pass the class
    rObject to the Initializer.
    * src/object/root-classes.cc: Create all classes, then initialize them.