Stop supporting "class a;", start supporting "class a : b {};".
It seems that "class a;" is not really required, but anyway it might
be possible to reactivate this if we really want it.

One nasty issue here is that the prescanner cuts everything when
it sees a comma alone, so "class a: b, c {}" is cut at the comma!
This will have to wait for the C++ push-parser.

	* src/parser/ast-factory.hh, src/parser/ast-factory.cc
	(ast_class): Accept an argument for the list of protos.
	* src/parser/utoken.l, src/parser/ugrammar.y ("public", "private")
	("protected"): Activate these keywords.
	(visibility, proto, protos, protos.1): new.
	(class instantiation): Stop supporting "class a;", start
	supporting "class a : b {};".