Loadable.
The k1 interface is too invasive on several regards.  The use of
"swich" is really a problem as it is the keyword for switch/case.
These routines were included for all, which is ugly, and prevented
the default definition of "load".

So now one has to derive from Loadable which brings "load", "on",
"off", and "toggle" (instead of switch).

IMHO, it is a pity to use "load", "on" and "off".  Especially the
first two.  Can't we rename "load" to something else?

	* share/urbi/k1.u: Move this code to...
	* share/urbi/loadable.u: this new file.
	* share/urbi/urbi.u: Adjust.
	* src/parser/utoken.l: switch is reserved.