Optimize visitors by using plain pointers.
In "operator()", "accept" and "visit", use plain pointers to AST
components instead of reference counted ones. As a consequence, we have
a 10% performance gain and a 33% to 50% stack space gain as the calls
from "accept" to "visit" are now tail calls.

It is possible to do so because when we are visiting, the visited node
is never destroyed as the visitor caller keeps a reference onto the
node. In the interpreter, if a node needs to be duplicated (because of a
"&" for example) a new shared pointer sharing the same intrusive counter
will be used.