Use eval() directly in the interpreter.
Sometimes, it is possible to use eval() instead of operator() in
the interpreter as we know that the underlying expression lies into
a scope that will then do the proper location handling in the case
of errors. This removes one stack level due to operator() not being
able to make tail calls.

	* src/runner/interpreter-visit.hxx: Use eval() instead of
	operator() whenever possible.