Do not desugar assigments and declarations directly.
The parser now simply creates Assignment and Declaration nodes. It is the role
of the binder to later rewrite them to LocalAssigment/LocalDeclaration nodes or
setSlots/updateSlots, depending on whether they are local or not.

This entails a better modularity (less work in the parser), and enables to
pretty print actual assignments instead of updateSlots. Moreover, this probably
enable to use lvalue meta variables for assignments in ParametricAsts.

    * src/ast/ast.yml
    (Write): New abstract node for writes (to local or slots).
    (Declaration): New node for local or slot declaration.
    (Assignment):  New node for local or slot assignment.
    * src/binder/binder.cc,
    * src/binder/binder.hh: Rewrite Assignments and Declaration to local writes
    or slots writes.
    * src/parser/parse.hh: Add forgotten include.
    * src/parser/ugrammar.y,
    * src/rewrite/pattern-rewriter.cc,
    * src/rewrite/rescoper.cc,
    * src/rewrite/rescoper.hh: Just create Assignment and Declaration nodes
    instead of Local ones.
    * src/runner/interpreter-visit.hxx: Reject Assignment and Declaration nodes.
    * tests: Update: adapt.