Migrate to a type safer interface of yylex.
Instead of using the old signature "int yylex(lvalue&, location&)", we
now use "symbol_type yylex()" and the "make_symbol" functions that ensure
a safer typing (it is not possible to return a STRING and set the int
value of the variant for instance).

As an unexpected side effect, we can no longer play dirty tricks
with two different signatures for yylex: one for the regular scanning, the
other for prescanning (we used to return forged integers, and now
we can only return declared tokens).  So we define new tokens just for the
prescanning.

	* src/parser/ugrammar.y (%define lex_symbol): New.
	(yylex): Fix the signature.
	(TOK_PRE_EOF, TOK_PRE_WANTS_MORE, TOK_PRE_COMPLETE): New, for
	prescanning only.
	* src/parser/flex-lexer.hh (yylex): Fix the signatures.
	Adjust the code.
	(grown_string, loc): New.
	We no longer use those of the look-ahead, the scanner is
	only using its members.
	* src/parser/parser-impl.hh (symbol_type): New.
	* src/parser/prescan.cc (prescan): Adjust to symbol_type.
	* src/parser/utoken.hh (yylex): Fix the signature.
	* src/parser/utoken.l (pre_eof, pre_wants_more, pre_complete):
	Remove.
	Adjust look-ahead constructions to use make_symbol.