Don't wrap \autoref.
Yes, it is a smart idea to factor the call to autoref and the type of the
entity, e.g., \sec{foo} instead of \autoref{sec:foo}'.  But (i) it breaks
bindings brought by AUC-TeX, (ii) it forces to use \sect because \sec is
already defined by LaTeX, and more importantly (iii) it makes it harder to
rename label since it is not the same sequence used to define a label, and
to use it.

	* doc/specs/urbi-specs.sty (\sect, \lst, \tab): Remove.
	* doc/specs/float.tex, doc/specs/list.tex, doc/specs/urbi-specs.tex:
	Adjust by using \autoref directly.
	Also, fix several incorrect labels.