Ship the generated parser.
We rely on a local Bison to generate our parser, but we don't include
a tarball of it in our K2 tarball (and I don't think we should).  So
generate the parser in the src tree, and ship it so that we don't have
dependencies on bison.  Fortunately the Bison tree was only needed when
generating the parser was needed.

	* src/parser/local.mk: Formatting changes.
	(parser_dir): Simplify.
	(FROM_UGRAMMAR_Y): Rename as...
	(SOURCES_FROM_UGRAMMAR_Y): this.
	(DATA_FROM_UGRAMMAR_Y): New.
	Ship and maintainerclean both of them.
	Adjust the rules.