Fix more distcheck issues.
As a thumb rule, one should never ever use "$(srcdir)" to denote files
in the source tree, even if the files are to be created in the src tree.
Rather, always use the short names for the files, and use $(srcdir)
where needed *in the rules* (i.e., in the shell commands), but not where
Make is involved.

Also, avoid using $< in that case, since it qualifies with $(srcdir) at
places we don't want to in general.

All this is quite complex because we also want to spare useless
compilations: even if calling Bison generates a fresh location.hh each
time, we do not want to recompile everything because of this, so we keep
location.hh in the past relatively to its stamp file...

	* src/object/local.mk: Filter out generated files.
	* src/parser/local.mk: Beware of $<.