Do not try to ship the SWIG generated files.
There is a bug Automake that made it ignore the variables (such as
liburbiPhp_la_SOURCES) that were defined *indented* in an if/endif.
When I unindented them, I activated their parsing by Automake who
started to honor them, in particular it wants to ship the sources.
But we don't want to.

	* src/swig/Makefile.am: Change all the *_SOURCES in
	nodist_*_SOURCES.