Replace the link + fork_point system by a simpler one.
The fork point mechanism obsoleted the link system between jobs,
but was too complicated and required code duplication in several
places (interpreter, "each&", and "load").

This new system allows a job to have a parent. If an exception
reaches the job toplevel and if the job has a parent, the exception
is reinjected into the parent, which must kill sibling jobs of the
job that just died and rethrow the exception.

An intermediate ChildException is used so that the reraised
exception cannot be mismatched by an exception arriving
during the parent regular control flow and caught by
mistake.

	* share/urbi/control.u (detach_linked): Remove, it was
	unused and is now obsolete.
	* src/scheduler/job.hh, src/scheduler/job.hxx,
	* src/scheduler/job.cc: Replace link system by parenting
	control.
	* src/runner/runner.hh, src/runner/runner.hxx,
	* src/runner/interpreter-visit.hxx, src/runner/interpreter.cc,
	* src/object/list.cc, src/object/system.cc: Use the new
	parenting system.
	* share/urbi/exception.u: Use "|" instead of ";" to please
	the tests.