Move scheduling errors handling from C++ to Urbi.
To restore the proper layering of the scheduler and object (the former
is not supposed to depend on the later), a new virtual method
"scheduling_error()" has been added, and is overriden in the interpreter.

Also, we no longer check that we are near stack space exhaustion if we
are generating the exception indicating that we have nearly exhausted
the stack space, so that we don't get infinitely recursive errors.

	* share/urbi/exception.u (SchedulingError): New.
	* src/object/urbi-exception.hh, src/object/urbi-exception.hxx
	(SchedulingError, StackExhaustedError): Remove.
	* src/scheduler/job.hh, src/scheduler/job.hxx,
	* src/scheduler/job.cc, src/scheduler/interpreter.hh,
	* src/scheduler/interpreter.cc (scheduling_error): New.
	* src/scheduler/tag.cc: Add missing include.