Fix memory leak in cloner.
The "recurse_collection()" function was always returning a pointer,
which may have been dereferenced in some cases but never freed.

Now, "recurse_collection()" always take a reference and return a
non-pointer object, and a new pointer is created when needed.

Also, "maybe_recurse_collection()" works with pointer and return 0
if given 0.

	* dev/ast-cloner-gen, src/ast/cloner.hxx (recurse_collection):
	Work with non-pointers.
	(maybe_recurse_collection): New.
	* src/ast/parametric-ast.cc, src/binder/binder.cc,
	* src/rewrite/desugarer.cc, src/rewrite/rescoper.cc: Use
	"maybe_recurse_collection" where appropriate.