Fix "isdef" and move it around.
We want to get rid of "isdef" as soon as possible, i.e. when we
have UObjects v2.  To make sure that we do not forget to remove this
abomination, move it to "uobject.u".

Also, split on dots and spaces, not "dot followed by space" since it
was wrong and didn't work. This is also more resilient to argString
format changes.

	* share/urbi/k1.u (isdef): Move from here to...
	* share/urbi/uobject.u (isdef): ...here after fixing.