Sanitize the interface of urbisend.
These changes are prompted by the need to be able to test these features
(UClient features run by the example programs).  This is on the way
towards installcheck too.

The code hurt my eyes, so I have also sanitized some of the badly chosen
interfaces.  Especially because this code is supposed to be released as
free software.

	* include/urbi/uabstractclient.hh, include/urbi/uclient.hh,
	* src/liburbi/uabstractclient.cc, src/liburbi/uclient.cc
	(host): Rename as...
	(host_): this, and make it a std::string.
	Adjust dependencies.
	(port): Rename as...
	(port_): this.
	Adjust dependencies.
	(sendFile): Use a std::string instead.
	* src/examples/urbisend.cc: Major revamp: support command line
	interface (including --help), and accept several files, not just
	one.