Fix plugin UObject naming.
When a new UObject is instanciated, an unique name is generated and
available to the user( as the __name variable). It can be used to do
stuffs like 'UVar v(__name+".stuff"); v=1;', and
'send("var " + __name +".stuff = 1;")'.
So, this name must be a valid identifier, and must be accessible in
urbi.

	* src/kernel/uobject.cc(uobject_new): Prefix the id identifying
	an object by 'uob_' to obtain a valid identifier. Put the
	object under this name where all the uobjects are.