Do not use raw scheduler tags in Urbi anymore.
The interpreters use two types of stacks: a stack of scheduler tags,
used for job control, and a stack of Urbi tags, used for manipulation
by the user. Flow control tags for example are not stored on the
Urbi tag stack, but only in the scheduler tags one.

This change removes the use of a scheduler tag stack to only keep
the Urbi one. When the user wants to see the tags, or when we
need to inherit them in another job, we filter out the flow control
tags.

This simplifies the implementation a lot and reduces the risk of
having desynchronized stacks.

	* src/runner/runner.hh, src/runner/runner.cc, src/runner/shell.cc,
	* src/runner/interpreter.hh, src/runner/interpreter-visit.hxx,
	* src/runner/interpreter.cc: Use only one tag stack.
	* src/runner/at-handler.cc: Remove the duplicate stack there
	as well.
	* src/object/system.cc, src/object/task.hh, src/object/task.cc:
	Update.
	* src/object/cxx-conversions.hxx: Remove unused conversion
	from and to "tags_type".